+++ main.tf
@@ -9,6 +9,7 @@ locals {
   centralized_policies   = try(local.model.sdwan.centralized_policies, {})
   configuration_groups   = try(local.model.sdwan.configuration_groups, {})
   feature_profiles       = try(local.model.sdwan.feature_profiles, {})
+  network_hierarchy      = try(local.model.sdwan.network_hierarchy, {})
   policy_groups          = try(local.model.sdwan.policy_groups, {})
   settings               = try(local.model.sdwan.settings, {})
   topology_groups        = try(local.model.sdwan.topology_groups, {})